requirements.txt
python-dotenv
pydantic
pandas
requests
pdfplumber
tenacity
tqdm
pytest
